fix: app hang on attach contact sometimes
Closes https://github.com/deltachat/deltachat-desktop/issues/5723.

It seems that sometimes this causes front-end to basically
infinitely spam the backend with `get_contacts_by_ids`.

`isContactLoaded` also starts returning `true`
as soon as we have started loading a contact,
and not only when we have loaded it.

This might not completely fix such a potential bug,
but this should at least make it not happen in practice.

This bug has likely been surfaced by
9bb605ccd6dc609f383b5e7d37e5df5dbe3500c6
(https://github.com/deltachat/deltachat-desktop/pull/5581).