perf: don't load chats, contacts, gallery twice
It seems that `react-window-infinite-loader` calls `isItemLoaded`
before React gets a chance to re-render with the new state of
e.g. `chatLoadState`, and so `react-window-infinite-loader`
calls `loadMoreItems` twice.

Let's convert this state to a ref then so that the changes to it
are reflected immediately.
It is fine to do so because apparently the `isItemLoaded` function
is not used for rendering, but only for determining whether
`loadMoreItems` needs to get called.

Related:
- https://github.com/deltachat/deltachat-desktop/pull/5725#pullrequestreview-3467878299.