fix: do not minimize the window before hiding it (#5774)
Minimizing the window with mainWindow.window.minimize()
before hiding it with mainWindow.window.hide()
results in a crash when mainWindow.window.show()
is called later when running on Wayland:
[3179:0628/212044.664566:FATAL:wayland_toplevel_window.cc(1123)] Should not be called with kMinimized state

I have tested the fix in `labwc` by clicking the icon
to hide and show the window multiple times,
as well as hiding the window manually by clicking
on the window name in the taskbar (wlr/taskbar module of waybar)
and then restoring it via the tray icon.
This also fixed the problem when running `niri`,
I cannot reproduce the crash with it as well.

Closes <https://github.com/deltachat/deltachat-desktop/issues/3989>