fix: openAlertDialog never resolving if Esc closed
And also do the same for `CopyContentAlertDialog`.

According to the search for `(?<!  )openAlertDialog\(`,
there are no places where this bug affects behavior
in a significant way.
This is mostly just about whether the dialog one layer below
(QR scanner) will get closed or will stay open.

Apart from `openAlertDialog`, there is another place
which relies on the `cb` argument, and that is `AccountSetupScreen`,
but there the behavior also seems unaffected.